requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
pika==1.2.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
pytest==6.2.4
pytest-asyncio==0.15.1
loguru==0.5.3
python-multipart==0.0.6